feat : 아이템 구조 개선
