--- Source/PullAndPush/Private/CollisionActionHandler.cpp
@@ -1,6 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "CollisionActionHandler.h"
-
-// Add default functionality here for any ICollisionActionHandler functions that are not pure virtual.
