+++ Source/PullAndPush/Private/Interface/AttackWeapon.cpp
@@ -1,0 +1,6 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Interface/AttackWeapon.h"
+
+// Add default functionality here for any IAttackWeapon functions that are not pure virtual.
