+++ Source/PullAndPush/Private/Interface/CollisionActionHandler.cpp
@@ -1,0 +1,6 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Interface/CollisionActionHandler.h"
+
+// Add default functionality here for any ICollisionActionHandler functions that are not pure virtual.
