--- Source/PullAndPush/Private/Item/InteractableItem.cpp
@@ -1,6 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "Item/InteractableItem.h"
-
-// Add default functionality here for any IInteractableItem functions that are not pure virtual.
