+++ Source/PullAndPush/Private/Item/Item.cpp
@@ -1,8 +1,13 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
-
 #include "Item/Item.h"
 
-UItem::UItem() {
+FString UItem::GetIdentifierString() const
+{
+	return GetPrimaryAssetId().ToString();
+}
 
+FPrimaryAssetId UItem::GetPrimaryAssetId() const
+{
+	return FPrimaryAssetId(ItemType, GetFName());
 }
