+++ Source/PullAndPush/Private/Item/ItemAssetManager.cpp
@@ -1,0 +1,37 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Item/ItemAssetManager.h"
+#include "Item/Item.h"
+
+const FPrimaryAssetType	UItemAssetManager::PassiveItemType = TEXT("Passive");
+const FPrimaryAssetType	UItemAssetManager::ActiveItemType = TEXT("Active");
+
+UItemAssetManager& UItemAssetManager::Get()
+{
+	UItemAssetManager* This = Cast<UItemAssetManager>(GEngine->AssetManager);
+
+	if (This)
+	{
+		return *This;
+	}
+	else
+	{
+		UE_LOG(LogTemp, Log, TEXT("Invalid AssetManager in DefaultEngine.ini, must be ItemAssetManager!"));
+		return *NewObject<UItemAssetManager>(); // never calls this
+	}
+}
+UItem* UItemAssetManager::ForceLoadItem(const FPrimaryAssetId& PrimaryAssetId, bool bLogWarning)
+{
+	FSoftObjectPath ItemPath = GetPrimaryAssetPath(PrimaryAssetId);
+
+	// This does a synchronous load and may hitch
+	UItem* LoadedItem = Cast<UItem>(ItemPath.TryLoad());
+
+	if (bLogWarning && LoadedItem == nullptr)
+	{
+		UE_LOG(LogTemp, Warning, TEXT("Failed to load item for identifier %s!"), *PrimaryAssetId.ToString());
+	}
+
+	return LoadedItem;
+}
