--- Source/PullAndPush/Private/RocketPunch/AttackWeapon.cpp
@@ -1,6 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "RocketPunch/AttackWeapon.h"
-
-// Add default functionality here for any IAttackWeapon functions that are not pure virtual.
