+++ Source/PullAndPush/Private/RocketPunch/RPCollisionComponent.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 #include "RocketPunch/RPCollisionComponent.h"
-#include "CollisionActionHandler.h"
+#include "Interface/CollisionActionHandler.h"
 
 URPCollisionComponent::URPCollisionComponent()
 	:
