--- Source/PullAndPush/Public/Item/InteractableItem.h
@@ -1,25 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "UObject/Interface.h"
-#include "InteractableItem.generated.h"
-
-// This class does not need to be modified.
-UINTERFACE(MinimalAPI)
-class UInteractableItem : public UInterface
-{
-	GENERATED_BODY()
-};
-
-/**
- * 
- */
-class PULLANDPUSH_API IInteractableItem
-{
-	GENERATED_BODY()
-
-	// Add interface functions to this class. This is the class that will be inherited to implement this interface.
-public:
-};
