+++ Source/PullAndPush/Public/Player/PlayableController.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/PlayerController.h"
+#include "PlayableController.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class PULLANDPUSH_API APlayableController : public APlayerController
+{
+	GENERATED_BODY()
+	
+};
